Rename GRPCRequest to GRPCMessage
